Refactored code and finished part 4.9
