+++ part4/backend/utils/logger.js
@@ -1,9 +1,13 @@
 const info = (...params) => {
-    console.log(params)
+    if(process.env.NODE_ENV !== 'test') {
+        console.log(...params)
+    }
 }
 
 const error = (...params) => {
-    console.log(params)
+    if(process.env.NODE_ENV !== 'test') {
+        console.log(...params)
+    }
 }
 
 module.exports = {
